Fixed rename, will try to clean up code
