+++ cms_frontend/src/types/FileFormat.tsx
@@ -1,0 +1,6 @@
+interface FileFormat {
+  filename: string,
+  type: string
+}
+
+export default FileFormat;
